Changes from feedback.
Moved bytesToHex to Utils class.
